+++ _tests/test/core/di/visibility_test.dart
@@ -1,12 +1,12 @@
 @TestOn('browser')
-
 import 'package:test/test.dart';
 import 'package:angular/angular.dart';
 import 'package:angular_test/angular_test.dart';
 
 import 'visibility_test.template.dart' as ng_generated;
 
-// TODO(leonsenft): expect specific DI error when introduced; b/64980526.
+final throwsNoProviderError = throwsA(const isInstanceOf<NoProviderError>());
+
 void main() {
   ng_generated.initReflector();
 
@@ -16,7 +16,7 @@ void main() {
     group('local', () {
       test('component should not be injectable by child component', () async {
         final testBed = new NgTestBed<ShouldFailToInjectParentComponent>();
-        expect(testBed.create(), throwsInAngular(anything));
+        expect(testBed.create(), throwsNoProviderError);
       });
 
       test('directive should be accessible via a query', () async {
@@ -29,17 +29,17 @@ void main() {
 
       test('directive should not be injectable on same element', () async {
         final testBed = new NgTestBed<ShouldFailToInjectFromElement>();
-        expect(testBed.create(), throwsInAngular(anything));
+        expect(testBed.create(), throwsNoProviderError);
       });
 
       test('directive should not be injectable in same view', () async {
         final testBed = new NgTestBed<ShouldFailToInjectFromView>();
-        expect(testBed.create(), throwsInAngular(anything));
+        expect(testBed.create(), throwsNoProviderError);
       });
 
       test('directive should not be injectable in child view', () async {
         final testBed = new NgTestBed<ShouldFailToInjectFromParentView>();
-        expect(testBed.create(), throwsInAngular(anything));
+        expect(testBed.create(), throwsNoProviderError);
       });
 
       test('service on Visibility.none component is injectable', () async {
@@ -81,6 +81,7 @@ void main() {
 )
 class InjectsVisibilityLocalComponent {
   ShouldFailToInjectParentComponent parent;
+
   InjectsVisibilityLocalComponent(this.parent);
 }
 
@@ -112,6 +113,7 @@ class ShouldQueryDirective {
 )
 class InjectsDirectiveComponent {
   VisibilityNoneDirective directive;
+
   InjectsDirectiveComponent(this.directive);
 }
 
@@ -187,6 +189,7 @@ class MyChildComponentProvidesService implements SomeService {
 )
 class MyDirectiveNeedsService {
   final SomeService someService;
+
   MyDirectiveNeedsService(
       this.someService, ViewContainerRef ref, TemplateRef templateRef);
 }
@@ -223,6 +226,7 @@ class InjectsAliasedLocal {
 )
 class InjectsVisibilityAllComponent {
   final ShouldInjectParentComponent parent;
+
   InjectsVisibilityAllComponent(this.parent);
 }
 
@@ -270,6 +274,7 @@ class VisibilityLocalImplementation implements Interface {}
 )
 class InjectsMultiToken {
   final List<Interface> dependencies;
+
   InjectsMultiToken(@implementations this.dependencies);
 }