+++ angular_ast/lib/src/ast/container.dart
@@ -15,12 +15,14 @@ const _listEquals = const ListEquality();
 /// Clients should not extend, implement, or mix-in this class.
 abstract class ContainerAst implements StandaloneTemplateAst {
   factory ContainerAst({
+    List<AnnotationAst> annotations,
     List<StandaloneTemplateAst> childNodes,
     List<StarAst> stars,
   }) = _SyntheticContainerAst;
 
   factory ContainerAst.from(
     TemplateAst origin, {
+    List<AnnotationAst> annotations,
     List<StandaloneTemplateAst> childNodes,
     List<StarAst> stars,
   }) = _SyntheticContainerAst.from;
@@ -29,9 +31,13 @@ abstract class ContainerAst implements StandaloneTemplateAst {
     SourceFile sourceFile,
     NgToken beginToken,
     NgToken endToken, {
+    List<AnnotationAst> annotations,
     List<StandaloneTemplateAst> childNodes,
     List<StarAst> stars,
   }) = _ParsedContainerAst;
+
+  /// Annotation assignments.
+  List<AnnotationAst> get annotations;
 
   /// Star assignments.
   List<StarAst> get stars;
@@ -81,9 +87,13 @@ class _ParsedContainerAst extends TemplateAst with ContainerAst {
     SourceFile sourceFile,
     NgToken beginToken,
     NgToken endToken, {
+    this.annotations: const [],
     this.childNodes: const [],
     this.stars: const [],
   }) : super.parsed(beginToken, endToken, sourceFile);
+
+  @override
+  final List<AnnotationAst> annotations;
 
   @override
   final List<StandaloneTemplateAst> childNodes;
@@ -94,15 +104,20 @@ class _ParsedContainerAst extends TemplateAst with ContainerAst {
 
 class _SyntheticContainerAst extends SyntheticTemplateAst with ContainerAst {
   _SyntheticContainerAst({
+    this.annotations: const [],
     this.childNodes: const [],
     this.stars: const [],
   });
 
   _SyntheticContainerAst.from(
     TemplateAst origin, {
+    this.annotations: const [],
     this.childNodes: const [],
     this.stars: const [],
   }) : super.from(origin);
+
+  @override
+  final List<AnnotationAst> annotations;
 
   @override
   final List<StandaloneTemplateAst> childNodes;