+++ angular_ast/test/parser_test.dart
@@ -229,6 +229,15 @@ void main() {
     );
   });
 
+  test('should handle a microsyntax expression with leading whitespace', () {
+    expect(
+      parse('<div *ngFor="\n  let item of items">{{item}}</div>'),
+      parse('<template ngFor let-item [ngForOf]="items">'
+          '<div>{{item}}</div>'
+          '</template>'),
+    );
+  });
+
   test('should parse a structural directive in child position', () {
     expect(
       parse('<div><div *ngIf="someValue">Hello World</div></div>'),
@@ -265,11 +274,37 @@ void main() {
     );
   });
 
-  test('should parse an annotation', () {
+  test('should parse and desugar @deferred', () {
     expect(parse('<div @deferred></div>'), [
       new EmbeddedTemplateAst(
           hasDeferredComponent: true,
           childNodes: [new ElementAst('div', new CloseElementAst('div'))])
+    ]);
+  });
+
+  test('should parse multiple annotations and desugar @deferred', () {
+    expect(parse('<div @foo="bar" @deferred></div>'), [
+      new EmbeddedTemplateAst(hasDeferredComponent: true, childNodes: [
+        new ElementAst('div', new CloseElementAst('div'), annotations: [
+          new AnnotationAst('foo', 'bar'),
+        ])
+      ]),
+    ]);
+  });
+
+  test('should parse an annotation with a value', () {
+    expect(parse('<div @foo="bar"></div>'), [
+      new ElementAst('div', new CloseElementAst('div'), annotations: [
+        new AnnotationAst('foo', 'bar'),
+      ]),
+    ]);
+  });
+
+  test('should parse an annotation on a container', () {
+    expect(parse('<ng-container @annotation></ng-container>'), [
+      new ContainerAst(annotations: [
+        new AnnotationAst('annotation'),
+      ])
     ]);
   });
 }