+++ angular_compiler/lib/src/cli/flags.dart
@@ -6,7 +6,7 @@ import 'package:meta/meta.dart';
 const _argDebugMode = 'debug';
 const _argProfileFor = 'profile';
 const _argLegacyStyle = 'use_legacy_style_encapsulation';
-const _argFastBoot = 'fast_boot';
+const _argLegacyWhitespace = 'use_legacy_preserve_whitespace';
 
 /// Compiler-wide configuration (flags) to allow opting in/out.
 ///
@@ -40,22 +40,19 @@ class CompilerFlags {
           'Whether to emit additional code that may be used by tooling '
           'in order to profile performance or other runtime information.',
     )
-    ..addFlag(_argFastBoot,
-        defaultsTo: null,
-        help: 'Whether to support AngularDart v5\'s "fastBoot" functionality.');
+    ..addFlag(
+      _argLegacyWhitespace,
+      defaultsTo: null,
+      help: 'Whether to opt-in/out of the new preserveWhitespace: false',
+      // It's not clear we will keep this flag through the 5.x final release.
+      hide: true,
+    );
 
   /// Whether to emit extra code suitable for testing and local development.
   final bool genDebugInfo;
 
   /// May emit extra code suitable for profiling or tooling.
   final Profile profileFor;
-
-  /// Whether to support AngularDart v5's "fastBoot" functionality.
-  ///
-  /// If `false`, additional code may be generated that negatively affects
-  /// startup performance and code size, but allows compatibility with
-  /// functionality such as `SlowComponentLoader` and `ReflectiveInjector`.
-  final bool useFastBoot;
 
   /// Whether to opt-in to supporting a legacy mode of style encapsulation.
   ///
@@ -84,7 +81,6 @@ class CompilerFlags {
     this.genDebugInfo: false,
     this.ignoreNgPlaceholderForGoldens: false,
     this.profileFor: Profile.none,
-    this.useFastBoot: true,
     this.useLegacyStyleEncapsulation: false,
     this.useNewPreserveWhitespace: false,
   });
@@ -130,7 +126,9 @@ class CompilerFlags {
         _argDebugMode,
         _argProfileFor,
         _argLegacyStyle,
-        _argFastBoot,
+        _argLegacyWhitespace,
+        // TODO(matanl): A better strategy for negatable options.
+        'no-$_argLegacyWhitespace',
       ].toSet();
       final unknownArgs = options.keys.toSet().difference(knownArgs);
       if (unknownArgs.isNotEmpty) {
@@ -158,17 +156,19 @@ class CompilerFlags {
       log('Invalid value for "$_argLegacyStyle": $useLegacyStyle');
       useLegacyStyle = null;
     }
-    var useFastBoot = options[_argFastBoot];
-    if (useFastBoot != null && useFastBoot is! bool) {
-      log('Invalid value for "$_argFastBoot": $useFastBoot');
-      useFastBoot = null;
+    var useLegacyWhitespace = options[_argLegacyWhitespace];
+    if (useLegacyWhitespace != null && useLegacyWhitespace is! bool) {
+      log('Invalid value for "$_argLegacyWhitespace": $useLegacyWhitespace');
+      useLegacyWhitespace = null;
     }
     return new CompilerFlags(
       genDebugInfo: debugMode ?? defaultTo.genDebugInfo,
       profileFor: _toProfile(profileFor, log) ?? defaultTo.profileFor,
       useLegacyStyleEncapsulation:
           useLegacyStyle ?? defaultTo.useLegacyStyleEncapsulation,
-      useFastBoot: useFastBoot ?? defaultTo.useFastBoot,
+      useNewPreserveWhitespace: useLegacyWhitespace != null
+          ? !useLegacyWhitespace
+          : defaultTo.useNewPreserveWhitespace,
     );
   }
 }