+++ angular_forms/CHANGELOG.md
@@ -1,7 +1,14 @@
+### New Features
+
+*   `Control.markAsPristine` added. This will clear the `dirty` property.
+
 ### Breaking Changes
 
 *   `NgControlName` will no longer initialize with `null` if a value is
     specified by 'ngModel'.
+
+*   The `touched` property of `Control`s is now propagated to parents /
+    children.
 
 ## 2.0.0-alpha+3